redux: move PersistenceRegistry to the redux feature
