+++ react/features/base/redux/PersistenceRegistry.js
