+++ react/features/base/redux/index.js
@@ -1,4 +1,5 @@
 export * from './functions';
 export { default as MiddlewareRegistry } from './MiddlewareRegistry';
+export { default as PersistenceRegistry } from './PersistenceRegistry';
 export { default as ReducerRegistry } from './ReducerRegistry';
 export { default as StateListenerRegistry } from './StateListenerRegistry';
