+++ react/features/base/settings/reducer.js
@@ -6,8 +6,7 @@ import _ from 'lodash';
 
 import { APP_WILL_MOUNT } from '../app';
 import { browser } from '../lib-jitsi-meet';
-import { ReducerRegistry } from '../redux';
-import { PersistenceRegistry } from '../storage';
+import { PersistenceRegistry, ReducerRegistry } from '../redux';
 import { assignIfDefined } from '../util';
 
 import { SETTINGS_UPDATED } from './actionTypes';
