--- react/features/base/storage/index.js
@@ -1,3 +1,0 @@
-export { default as PersistenceRegistry } from './PersistenceRegistry';
-
-import './middleware';
