--- react/features/base/storage/logger.js
@@ -1,5 +1,0 @@
-// @flow
-
-import { getLogger } from '../logging/functions';
-
-export default getLogger('features/base/storage');
