+++ react/features/dropbox/reducer.js
@@ -1,7 +1,6 @@
 // @flow
 
-import { ReducerRegistry } from '../base/redux';
-import { PersistenceRegistry } from '../base/storage';
+import { PersistenceRegistry, ReducerRegistry } from '../base/redux';
 
 import { UPDATE_DROPBOX_TOKEN } from './actionTypes';
