+++ react/features/recent-list/reducer.js
@@ -4,8 +4,7 @@ import { jitsiLocalStorage } from 'js-utils';
 
 import { APP_WILL_MOUNT } from '../base/app';
 import { getURLWithoutParamsNormalized } from '../base/connection';
-import { ReducerRegistry } from '../base/redux';
-import { PersistenceRegistry } from '../base/storage';
+import { PersistenceRegistry, ReducerRegistry } from '../base/redux';
 
 import {
     _STORE_CURRENT_CONFERENCE,
